requirements.txt
# ===================================================================
# solstice-core requirements file
# ===================================================================
Qt.py
sentry-sdk==0.12.2
tpPyUtils
tpDccLib
artellapipe-utils
artellapipe-core
artellapipe-gui
